# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Employee & Auth API
  version: 1.0.0
  description: API for managing employees and authentication
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
paths:
  /employee:
    get:
      tags:
        - Employee
      summary: Get paginated, filtered, and sorted list of employees
      description: Returns a paginated list of employees with optional occupation filter and sorting.
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number (default is 1)

        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of employees per page

        - in: query
          name: occupation
          schema:
            type: string
            enum:
              - developer
              - sales
              - hr
              - management
              - customer_success
              - tester
          description: Filter by employee occupation

        - in: query
          name: sortBy
          schema:
            type: string
            enum:
              - firstname
              - lastname
              - age
              - occupation
              - salary
              - employed
          description: Field to sort by

        - in: query
          name: sortDir
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort direction (ascending or descending)

      responses:
        '200':
          description: Paginated and filtered list of employees
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 10
                  total:
                    type: integer
                    example: 35
                  totalPages:
                    type: integer
                    example: 4
                  employees:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        firstname:
                          type: string
                          example: Alice
                        lastname:
                          type: string
                          example: Smith
                        age:
                          type: integer
                          example: 28
                        occupation:
                          type: string
                          enum:
                            - developer
                            - sales
                            - hr
                            - management
                            - customer_success
                            - tester
                          example: developer
                        salary:
                          type: number
                          example: 55000
                        employed:
                          type: string
                          format: date
                          example: 2022-05-10

        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid pagination parameters

        '401':
          description: Unauthorized, token missing or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No token provided

        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
    post:
      tags:
        - Employee
      summary: Create a new employee
      description: Creates a new employee with full details.
      security:
        - cookieAuth: []
      requestBody:
        description: Employee data
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstname
                - lastname
                - age
                - occupation
                - salary
                - employed
              properties:
                firstname:
                  type: string
                lastname:
                  type: string
                age:
                  type: integer
                occupation:
                  type: string
                  enum:
                    - developer
                    - sales
                    - hr
                    - management
                    - customer_success
                    - tester
                salary:
                  type: number
                employed:
                  type: string
                  format: date
              example:
                firstname: John
                lastname: Doe
                age: 30
                occupation: developer
                salary: 60000
                employed: 2024-01-15
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  firstname:
                    type: string
                    example: John
                  lastname:
                    type: string
                    example: Doe
                  age:
                    type: integer
                    example: 30
                  occupation:
                    type: string
                    example: developer
                  salary:
                    type: number
                    example: 60000
                  employed:
                    type: string
                    format: date
                    example: 2024-01-15
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Missing required data to create employee
        '401':
          description: Unauthorized, token missing or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No token provided
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error

    delete:
      tags:
        - Employee
      summary: Delete multiple employees
      description: Delete one or more employees by ID.
      security:
        - cookieAuth: []
      requestBody:
        description: Array of employee IDs to delete
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
              example:
                ids: ['1', '2', '3']
      responses:
        '200':
          description: Employees deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                    example: 3
        '400':
          description: No employee IDs provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No employee IDs were provided
        '401':
          description: Unauthorized, token missing or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No token provided
        '404':
          description: No matching employees found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No matching employees found
        '500':
          description: Internal server error

  /employee/{id}:
    get:
      tags:
        - Employee
      summary: Get employee by ID
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Employee ID
      responses:
        '200':
          description: Employee found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  firstname:
                    type: string
                    example: Alice
                  lastname:
                    type: string
                    example: Smith
                  age:
                    type: integer
                    example: 28
                  occupation:
                    type: string
                    example: developer
                  salary:
                    type: number
                    example: 55000
                  employed:
                    type: string
                    format: date
                    example: 2022-05-10
        '400':
          description: Missing or invalid ID
        '401':
          description: Unauthorized, token missing or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No token provided
        '404':
          description: Employee not found
        '500':
          description: Internal server error

    put:
      tags:
        - Employee
      summary: Update employee by ID
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Employee ID
      requestBody:
        description: Updated employee data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                lastname:
                  type: string
                age:
                  type: integer
                occupation:
                  type: string
                  enum:
                    - developer
                    - sales
                    - hr
                    - management
                    - customer_success
                    - tester
                salary:
                  type: number
                employed:
                  type: string
                  format: date
              example:
                firstname: John
                lastname: Smith
                age: 35
                occupation: developer
                salary: 70000
                employed: 2024-01-01
      responses:
        '200':
          description: Employee updated successfully
        '400':
          description: Missing or invalid data
        '401':
          description: Unauthorized, token missing or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No token provided
        '404':
          description: Employee not found
        '500':
          description: Internal server error

  /auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        description: Login credentials
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: johndoe
                password: secret
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                example:
                  'success': true,
                  'message': 'Login successful'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid username or password

  /auth/register:
    post:
      tags:
        - Auth
      summary: User registration
      description: Register a new user
      requestBody:
        description: User registration info
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: johndoe
                password: secret
      responses:
        '201':
          description: Registration successful
        '400':
          description: User already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User already exists
        '500':
          description: Internal server error